break up some of the larger code presentation-only
